+++ tests/api/one/mod.rs
@@ -1,1 +1,1 @@
-pub mod accounts;
+pub mod accounts_test;
